[⚙️ chore] 챌린지 리뷰 컴포넌트 분리 #89
